requirements.txt
python-dotenv
openai
pyupbit
streamlit
requests
plotly
schedule